requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.10.2
python-multipart==0.0.12
sqlalchemy==2.0.36
alembic==1.14.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing dependencies
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.28.1
pytest-cov==5.0.0
factory-boy==3.3.1
pytest-mock==3.14.0
PyJWT==2.10.1